requirements.txt
biopython==1.77
pandas==1.1.4
XlsxWriter==1.3.7
xlrd==1.2.0